+++ src/backend/domain/model/schedule/availableTime.ts
@@ -1,0 +1,4 @@
+export type AvailableTime = {
+  startTime: Date
+  endTime: Date
+}
