+++ src/backend/domain/model/time_slot/timeSlot.ts
